use of xxxScale enums reviewed Avg implement in sensor3D
Accelerometer scale error (4G) corrected to 2G

+++ RPISensors/src/main/java/sensors/Implementations/MPU9250/MPU9250Constants.java
@@ -221,9 +221,9 @@ enum AccScale
     {
         return (byte)value;
     }
-    public double getRes()
+    public float getRes()
     {
-        return (double)minMax/32768.0;
+        return minMax/32768.0f;
     }
     public int getMinMax()
     {
@@ -250,9 +250,9 @@ enum GyrScale
     {
         return (byte)value;
     }
-    public double getRes()
+    public float getRes()
     {
-        return (double)minMax/32768.0;
+        return minMax/32768.0f;
     }
     public int getMinMax()
     {
